src/third_party/blink/web_tests/virtual/stable/http/tests/origin_trials/feature-policy-reporting/feature-policy-reporting-no-trial-expected.txt

CONSOLE WARNING: line 21: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 8: Unable to preventDefault inside passive event listener due to target being treated as passive. See https://www.chromestatus.com/features/5093566007214080
This is a testharness.js-based test.
PASS Feature policy report only mode without the corresponding origin trial
Harness: the test ran to completion.
